Move existing resources to core::utils
- Moved SignalManager
- Moved getInstallPrefix
- New generic logger
