+++ cpp/scenario/core/src/signals.cpp
@@ -2,40 +2,22 @@
  * Copyright (C) 2020 Istituto Italiano di Tecnologia (IIT)
  * All rights reserved.
  *
- * This project is dual licensed under LGPL v2.1+ or Apache License.
- *
- * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
- *
  * This software may be modified and distributed under the terms of the
  * GNU Lesser General Public License v2.1 or any later version.
- *
- * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * * *
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
  */
 
-#include "scenario/gazebo/signals.h"
-#include "scenario/gazebo/Log.h"
+#include "scenario/core/utils/signals.h"
+#include "scenario/core/utils/Log.h"
 
 #include <csignal>
 #include <mutex>
 #include <unordered_map>
 
-namespace scenario::base::detail {
+namespace scenario::core::detail {
     static std::mutex SignalManagerMutex;
 }
 
-using namespace scenario::base;
+using namespace scenario::core::utils;
 
 class SignalManager::Impl
 {
